e3-causes-of-overhead/e3.1-cprofile/requirements.txt
numpy==2.0.2
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
jinja2==3.1.4
python-igraph==0.11.6